refactor(riviere-extract-conventions): convert ESLint plugin to CommonJS format
Switch from ESM to CommonJS for ESLint plugin to ensure compatibility with
Node.js require() usage. ESLint plugins typically expect CommonJS format.

Changes:
- Convert index.js and require-component-decorator.js to .cjs files
- Update type declarations to use .d.ts and .d.cts for CommonJS compatibility
- Update package.json exports to point to .cjs files
- Update package.json files array to include .js and .d.ts patterns
