+++ packages/riviere-extract-conventions/src/eslint/index.cjs
@@ -1,3 +1,3 @@
-const requireComponentDecorator = require('./require-component-decorator')
+const requireComponentDecorator = require('./require-component-decorator.cjs')
 
 module.exports = { rules: { 'require-component-decorator': requireComponentDecorator } }
