+++ packages/riviere-extract-conventions/src/eslint/require-component-decorator.cjs
