+++ packages/riviere-extract-conventions/src/eslint/require-component-decorator.d.cts
@@ -1,0 +1,4 @@
+import type { TSESLint } from '@typescript-eslint/utils'
+
+declare const rule: TSESLint.RuleModule<'missingDecorator'>
+export default rule
