--- packages/riviere-extract-conventions/src/eslint/require-component-decorator.d.ts
@@ -1,4 +1,0 @@
-import type { RuleModule } from '@typescript-eslint/rule-tester'
-
-declare const rule: RuleModule<'missingDecorator'>
-export default rule
